Fix testsuites
When we support ilp32 ABI on rv64 ISA, long int type wound generate *w instructions.
Change these testsuites and make it clear that they should use LP64.

gcc/testsuite/ChangeLog:

        * gcc.dg/lower-subreg-1.c: Skip if rv64 ISA.
        * gcc.dg/lto/pr88077_0.c: Likewise.
        * gcc.dg/pr90838.c: Add -mabi=lp64d.
        * gcc.dg/tree-ssa/ssa-dom-cse-2.c:Add -mabi=lp64d.
        * gcc.dg/unroll-7.c: Drop number iterations.
        * gcc.target/riscv/xtheadba-addsl.c:Add -mabi=lp64d.
        * gcc.target/riscv/xtheadbb-rev.c:Add -mabi=lp64d.
        * gcc.target/riscv/xtheadbb-srri.c:Add -mabi=lp64d.
        * gcc.target/riscv/xtheadmac-mula-muls.c:Add -mabi=lp64d.
        * gcc.target/riscv/xtheadmempair-3.c:Add -mabi=lp64d.

+++ gcc/testsuite/gcc.dg/lto/pr88077_0.c
@@ -1,3 +1,3 @@
 /* { dg-lto-do link } */
 
-long HeaderStr;
+long long HeaderStr;
